server/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==1.0.1
cytoolz==0.12.3
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.2.2
eth-utils==4.1.0
eth_abi==5.1.0
frozenlist==1.4.1
gunicorn==21.2.0
hexbytes==0.3.1
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
multidict==6.0.5
packaging==23.2
parsimonious==0.10.0
pillow==10.3.0
protobuf==5.26.1
psycopg2-binary==2.9.9
pycryptodome==3.20.0
PyJWT==2.8.0
pytz==2024.1
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rlp==4.0.1
rpds-py==0.18.0
sqlparse==0.4.4
toolz==0.12.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
web3==6.18.0
websockets==12.0
yarl==1.9.4
